refactor: atom 파일 import 수정
